+++ crates/builder-api/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "essential-builder-api"
 description = "API implementation for the Essential builder"
-version = "0.1.0"
+version = "0.2.0"
 authors.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -11,6 +11,7 @@ repository.workspace = true
 [dependencies]
 axum = { workspace = true }
 essential-builder-db = { workspace = true }
+essential-builder-types = { workspace = true }
 essential-types = { workspace = true }
 futures = { workspace = true }
 hex = { workspace = true }
