+++ crates/builder-db/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "essential-builder-db"
 description = "The Essential builder database"
-version = "0.1.0"
+version = "0.2.0"
 authors.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -9,6 +9,7 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
+essential-builder-types = { workspace = true }
 essential-hash = { workspace = true }
 essential-types = { workspace = true }
 num_cpus = { workspace = true, optional = true }
