+++ crates/builder-db/src/lib.rs
@@ -19,6 +19,7 @@
 //!   stored number is within a given limit.
 
 use error::{DecodeError, QueryError};
+use essential_builder_types::SolutionFailure;
 use essential_hash::content_addr;
 use essential_types::{solution::Solution, ContentAddress, Hash};
 #[cfg(feature = "pool")]
@@ -31,22 +32,6 @@ pub mod error;
 #[cfg(feature = "pool")]
 pub mod pool;
 pub mod sql;
-
-/// A failed attempt at applying a solution at a particular location within a particular block.
-///
-/// The builder stores these in order to provide solution submitters feedback in the case that a
-/// solution could not be applied trivially and did not make it into a block.
-#[derive(Clone, Debug, Eq, Hash, PartialEq, PartialOrd, Ord, Deserialize, Serialize)]
-pub struct SolutionFailure<'a> {
-    /// The number of the block in which the builder attempted to apply the solution.
-    pub attempt_block_num: i64,
-    /// The address of the block in which the builder attempted to apply the solution.
-    pub attempt_block_addr: ContentAddress,
-    /// The solution index within the block at which the builder attempted to apply the solution.
-    pub attempt_solution_ix: u32,
-    /// An error message describing why the builder failed to apply the solution.
-    pub err_msg: std::borrow::Cow<'a, str>,
-}
 
 /// Encodes the given value into a blob.
 ///
