+++ crates/builder/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "essential-builder"
 description = "A block builder library implementation for the Essential protocol"
-version = "0.1.0"
+version = "0.2.0"
 authors.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -11,6 +11,7 @@ repository.workspace = true
 [dependencies]
 essential-check = { workspace = true }
 essential-builder-db = { workspace = true }
+essential-builder-types = { workspace = true }
 essential-hash = { workspace = true }
 essential-node = { workspace = true }
 essential-node-db = { workspace = true }
